refactor: 스크린사이즈 key 타입 지정
