+++ src/components/atoms/text/Text.stories.ts
@@ -1,6 +1,5 @@
 import { Meta, StoryObj } from '@storybook/react';
 import Text from './Text';
-import { BREAKPOINT_MD } from '@styles/breakpoints';
 
 const meta: Meta<typeof Text> = {
   component: Text,
@@ -16,7 +15,7 @@ export const Heading: Story = {
     children: 'Heading 테스트 입니다.',
     defaultSize: '32px',
     responsiveSize: {
-      [BREAKPOINT_MD]: '46px',
+      md: '46px',
     },
     weight: 700,
   },
