+++ src/components/atoms/text/TextFactory.tsx
@@ -1,6 +1,5 @@
-import { TextProps } from '@/types/atoms';
+import { TextProps } from '@/types/props';
 import Text from '@components/atoms/text/Text';
-import { BREAKPOINT_MD } from '@styles/breakpoints';
 
 export function createTextComponent(props: Partial<TextProps>) {
   return function TextComponent(extraProps: TextProps) {
@@ -14,21 +13,21 @@ export const TextBody = {
   Large: createTextComponent({
     defaultSize: '15px',
     responsiveSize: {
-      [BREAKPOINT_MD]: '22px',
+      md: '22px',
     },
     as: 'p',
   }),
   Medium: createTextComponent({
     defaultSize: '13px',
     responsiveSize: {
-      [BREAKPOINT_MD]: '18px',
+      md: '18px',
     },
     as: 'p',
   }),
   Small: createTextComponent({
     defaultSize: '11px',
     responsiveSize: {
-      [BREAKPOINT_MD]: '16px',
+      md: '16px',
     },
     as: 'p',
   }),
@@ -38,7 +37,7 @@ export const Heading = {
   XLarge: createTextComponent({
     defaultSize: '32px',
     responsiveSize: {
-      [BREAKPOINT_MD]: '46px',
+      md: '46px',
     },
     as: 'h1',
     weight: 'bold',
@@ -46,7 +45,7 @@ export const Heading = {
   Large: createTextComponent({
     defaultSize: '28px',
     responsiveSize: {
-      [BREAKPOINT_MD]: '40px',
+      md: '40px',
     },
     as: 'h2',
     weight: 'bold',
@@ -54,7 +53,7 @@ export const Heading = {
   Medium: createTextComponent({
     defaultSize: '24px',
     responsiveSize: {
-      [BREAKPOINT_MD]: '34px',
+      md: '34px',
     },
     as: 'h3',
     weight: 'bold',
@@ -62,7 +61,7 @@ export const Heading = {
   Small: createTextComponent({
     defaultSize: '20px',
     responsiveSize: {
-      [BREAKPOINT_MD]: '28px',
+      md: '28px',
     },
     as: 'h4',
     weight: 'bold',
@@ -70,7 +69,7 @@ export const Heading = {
   XSmall: createTextComponent({
     defaultSize: '18px',
     responsiveSize: {
-      [BREAKPOINT_MD]: '26px',
+      md: '26px',
     },
     as: 'h5',
     weight: 'bold',
@@ -81,21 +80,21 @@ export const Paragraph = {
   Large: createTextComponent({
     defaultSize: '18px',
     responsiveSize: {
-      [BREAKPOINT_MD]: '30px',
+      md: '30px',
     },
     as: 'p',
   }),
   Medium: createTextComponent({
     defaultSize: '16px',
     responsiveSize: {
-      [BREAKPOINT_MD]: '26px',
+      md: '26px',
     },
     as: 'p',
   }),
   Small: createTextComponent({
     defaultSize: '14px',
     responsiveSize: {
-      [BREAKPOINT_MD]: '24px',
+      md: '24px',
     },
     as: 'p',
   }),
